+++ temba/orgs/models.py
@@ -3,6 +3,7 @@ import logging
 import os
 from abc import ABCMeta
 from collections import defaultdict
+from datetime import timedelta
 from enum import Enum
 from urllib.parse import quote, urlencode, urlparse
 
@@ -1240,12 +1241,12 @@ class Org(SmartModel):
 
     def release(self, user, *, release_users=True):
         """
-        Releases this org, marking it as inactive. Actual deletion of org data won't happen until after 7 days unless
-        delete is True.
+        Releases this org, marking it as inactive. Actual deletion of org data won't happen until after 7 days.
         """
 
-        # free our children
-        Org.objects.filter(parent=self).update(parent=None)
+        # release any child orgs
+        for child in self.children.all():
+            child.release(user, release_users=release_users)
 
         # deactivate ourselves
         self.is_active = False
@@ -1274,18 +1275,23 @@ class Org(SmartModel):
         Does an actual delete of this org
         """
 
-        assert not self.is_active and self.released_on, "can't delete an org which hasn't been released"
-        assert not self.deleted_on, "can't delete an org twice"
+        assert not self.is_active and self.released_on, "can't delete org which hasn't been released"
+        assert self.released_on < timezone.now() - timedelta(days=7), "can't delete org which was released recently"
+        assert not self.deleted_on, "can't delete org twice"
 
         user = self.modified_by
 
         # delete notifications and exports
         self.incidents.all().delete()
         self.notifications.all().delete()
+        self.notification_counts.all().delete()
         self.exportcontactstasks.all().delete()
         self.exportmessagestasks.all().delete()
         self.exportflowresultstasks.all().delete()
         self.exportticketstasks.all().delete()
+
+        for imp in self.contact_imports.all():
+            imp.delete()
 
         for label in self.msgs_labels.all():
             label.release(user)
@@ -1320,6 +1326,7 @@ class Org(SmartModel):
         self.sessions.all().delete()
         self.ticket_events.all().delete()
         self.tickets.all().delete()
+        self.ticket_counts.all().delete()
         self.topics.all().delete()
         self.airtime_transfers.all().delete()
 
@@ -1332,9 +1339,8 @@ class Org(SmartModel):
         self.urns.all().delete()
 
         # delete our fields
-        for contactfield in self.fields.all():
-            contactfield.release(user)
-            contactfield.delete()
+        for field in self.fields.all():
+            field.delete()
 
         # delete our groups
         for group in self.groups.all():
@@ -1349,15 +1355,13 @@ class Org(SmartModel):
 
             channel.delete()
 
-        for g in self.globals.all():
-            g.release(user)
+        for glob in self.globals.all():
+            glob.delete()
 
-        # delete our classifiers
         for classifier in self.classifiers.all():
             classifier.release(user)
             classifier.delete()
 
-        # delete our ticketers
         for ticketer in self.ticketers.all():
             ticketer.release(user)
             ticketer.delete()
@@ -1407,6 +1411,9 @@ class Org(SmartModel):
             "default_country": self.default_country_code,
             "redaction_policy": "urns" if self.is_anon else "none",
         }
+
+    def __repr__(self):
+        return f'<Org: name="{self.name}">'
 
     def __str__(self):
         return self.name