+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v8.1.79 (2023-03-20)
+-------------------------
+ * Add mouseover text for temba-date
+ * Reload page on org mismatch
+ * Use embedded title instead of response header
+
+v8.1.78 (2023-03-20)
+-------------------------
+ * Add globals to new ui
+ * Make it harder to accidentally delete an org
+ * Rewrite org deletion test and fix deletion issues
+
 v8.1.77 (2023-03-16)
 -------------------------
  * Limit groups to a single line on contact page